+++ msPCA.Rcheck/00_pkg_src/msPCA/src/RcppExports.cpp
@@ -12,8 +12,8 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // iterativeDeflationHeuristic
-List iterativeDeflationHeuristic(Eigen::MatrixXd Sigma, int r, Rcpp::NumericVector ks, int maxIter, bool verbose, int feasibilityConstraintType, double feasibilityTolerance, double stallingTolerance, int maxIterTPW, int timeLimitTPW);
-RcppExport SEXP _msPCA_iterativeDeflationHeuristic(SEXP SigmaSEXP, SEXP rSEXP, SEXP ksSEXP, SEXP maxIterSEXP, SEXP verboseSEXP, SEXP feasibilityConstraintTypeSEXP, SEXP feasibilityToleranceSEXP, SEXP stallingToleranceSEXP, SEXP maxIterTPWSEXP, SEXP timeLimitTPWSEXP) {
+List iterativeDeflationHeuristic(Eigen::MatrixXd Sigma, int r, Rcpp::NumericVector ks, int maxIter, bool verbose, int feasibilityConstraintType, double feasibilityTolerance, double stallingTolerance, int maxIterTPM, int timeLimitTPM, int restartsAfterFirstIter);
+RcppExport SEXP _msPCA_iterativeDeflationHeuristic(SEXP SigmaSEXP, SEXP rSEXP, SEXP ksSEXP, SEXP maxIterSEXP, SEXP verboseSEXP, SEXP feasibilityConstraintTypeSEXP, SEXP feasibilityToleranceSEXP, SEXP stallingToleranceSEXP, SEXP maxIterTPMSEXP, SEXP timeLimitTPMSEXP, SEXP restartsAfterFirstIterSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -25,9 +25,10 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type feasibilityConstraintType(feasibilityConstraintTypeSEXP);
     Rcpp::traits::input_parameter< double >::type feasibilityTolerance(feasibilityToleranceSEXP);
     Rcpp::traits::input_parameter< double >::type stallingTolerance(stallingToleranceSEXP);
-    Rcpp::traits::input_parameter< int >::type maxIterTPW(maxIterTPWSEXP);
-    Rcpp::traits::input_parameter< int >::type timeLimitTPW(timeLimitTPWSEXP);
-    rcpp_result_gen = Rcpp::wrap(iterativeDeflationHeuristic(Sigma, r, ks, maxIter, verbose, feasibilityConstraintType, feasibilityTolerance, stallingTolerance, maxIterTPW, timeLimitTPW));
+    Rcpp::traits::input_parameter< int >::type maxIterTPM(maxIterTPMSEXP);
+    Rcpp::traits::input_parameter< int >::type timeLimitTPM(timeLimitTPMSEXP);
+    Rcpp::traits::input_parameter< int >::type restartsAfterFirstIter(restartsAfterFirstIterSEXP);
+    rcpp_result_gen = Rcpp::wrap(iterativeDeflationHeuristic(Sigma, r, ks, maxIter, verbose, feasibilityConstraintType, feasibilityTolerance, stallingTolerance, maxIterTPM, timeLimitTPM, restartsAfterFirstIter));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -48,7 +49,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_msPCA_iterativeDeflationHeuristic", (DL_FUNC) &_msPCA_iterativeDeflationHeuristic, 10},
+    {"_msPCA_iterativeDeflationHeuristic", (DL_FUNC) &_msPCA_iterativeDeflationHeuristic, 11},
     {"_msPCA_truncatedPowerMethod", (DL_FUNC) &_msPCA_truncatedPowerMethod, 5},
     {NULL, NULL, 0}
 };