+++ msPCA.Rcheck/00_pkg_src/msPCA/NEWS.md
@@ -1,3 +1,10 @@
+# msPCA 0.3.0
+
+* Improved scalability of `mspca()` and `tpw()` through algorithmic and implementation optimizations.
+* Function `mspca()` now accepts a new hyper-parameter `restartsAfterFirstIter` that limits the number of random restarts for the truncated power method after the first outer iteration
+* Improved scaling of the penalty parameters for the case of zero-correlation constraints
+* Fixed incorrect acronym for truncated power method (TPW <- TPM) 
+
 # msPCA 0.2.0
 
 * Added support for no-correlation constraints between PCs as well as orthogonality constraints. User chooses between orthogonality and uncorrelatedness constraints via the `feasibilityConstraintType` parameter to `msPCA()`.